Implemented multiple speakers and speaker focusing!
Note: This is only in the scene named "DialogueUpdateScene.unity". The
next to do is to find a way to merge this into the DiaSystem prefab.

Usage of new/updated yarn commands:
<<ChangeSpeaker P1 P2>>
- dialogue with two speakers: P1 and P2.

<<Focus P1>>
- Focus on P1 (this requires "ChangeSpeaker P1" to have been called)

<<Focus FIRST|SECOND|BOTH|NONE>>
- Focus on the speaker on the left|speaker on the right|both
speakers|neither speaker.

ChangeSpeaker.cs
- Added another command here, called Focus. The main reason it was added
here was because of how much it would be needed with ChangeSpeaker.cs

DialogueUpdateScene.unity, temp.yarn
- contains the test dialogue